Tested with localhost API & files rearranged for cleaner code
